Crate proxy_attestation_client_bin

Client for the Proxy Attestation example.

Client first connects to the Proxy Attestation Service and downloads its root certificate. Then client uses this certificate in order to connect to the example Oak application, assuming that the example application received a signed certificate from the Proxy Attestation Service.

Structs

Opt

Constants

EXAMPLE_MESSAGE

INVALID_TEST_TEE_MEASUREMENT

TEST_TEE_MEASUREMENT

Functions

create_application_client

Create gRPC client for Oak application.

create_proxy_client

Create gRPC client for Proxy Attestation Service.

get_example_message

Connect to Oak application using root TLS certificate received from Proxy Attestation Service.

get_root_tls_certificate

Request root TLS certificate from Proxy Attestation Service.

main